+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/TestUtils.kt
@@ -422,7 +422,7 @@ fun Class<*>.createGraphWithNoArgs(): Any {
 fun Class<*>.createGraphViaFactory(vararg args: Any): Any {
   val factoryInstance = invokeGraphFactory()
   return factoryInstance.javaClass.declaredMethods
-    .single { it.name == Symbols.StringNames.CREATE }
+    .single { it.name == Symbols.StringNames.CREATE && Modifier.isFinal(it.modifiers) }
     .invoke(factoryInstance, *args)
 }